requirements.txt
django~=3.2.9
pillow
psycopg2-binary

django-haystack~=3.1.1
djangorestframework~=3.13.1